Improvements, Bug fixes
Clicking on a cage that contains a Dragon will show the Dragon's name and personality which are both chosen randomly, and the income the Dragon will make, which is based off the personality (Dragons do not actually generate gold yet, this will be implemented in the future).
Added camera movement status text.
Added the ability to cancel what you are placing in the shop.

Several bug fixes